Update bignumber.js to 5.0.0
This is a major version bump, but the changelog only has one change:
"Don't throw on constructor call without `new`."
This is not really even a breaking change.

There are newer versions of this library, but with actual breaking
changes. I wanted to balance risk with getting a more recent, fuller version of
its types.